add and implement stock interface to product service

+++ stock-service/src/main/java/edu/miu/stockservice/repository/StockRepository.java
@@ -4,9 +4,10 @@ import edu.miu.stockservice.entity.Stock;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 @Repository
-public interface StockRepository extends MongoRepository<Stock,Long> {
+public interface StockRepository extends MongoRepository<Stock, BigInteger> {
 
     Optional<Stock> findByProductNumber(String productNumber);
